Fix #80
