+++ app/modules/transcribe/overlay/graphic-dialog.html
@@ -1,7 +1,7 @@
 <div aria-label="Graphic" aria-hidden="{{ !active }}" class="graphic-dialog" ng-class="{ '-active': vm.active }" ng-style="position" role="dialog" tabindex="-1">
 
     <div class="heading">
-        <h3 class="title">Select one of the following graphic tags</h3>
+        <h3 class="title">Select one graphic type</h3>
         <button class="close" role="button" ng-click="vm.close($event)">
             <span class="icon" aria-hidden="true"></span>
             <span class="text">Close</span>
