--- app/modules/transcribe/overlay/keypad-slide.directive.js
@@ -1,59 +1,0 @@
-'use strict';
-
-var angular = require('angular');
-
-require('./overlay.module.js')
-    .directive('keypadSlide', keypadSlide);
-
-// @ngInject
-function keypadSlide(hotkeys, overlayConfig) {
-    var directive = {
-        link: keypadSlideLink,
-        controller: ['$scope', '$element', '$sce', keypadSlideController],
-        replace: true,
-        scope: true,
-        templateUrl: 'overlay/keypad-slide.html'
-    };
-    return directive;
-
-    function keypadSlideController($scope, $element, $sce) {
-        $scope.abbreviations = overlayConfig.abbrKeys;
-        $scope.tags = overlayConfig.teiTags;
-        var textarea = angular.element('textarea').first();
-
-        $scope.addTag = function (tagText) {
-            var startTag = '<' + tagText + '>';
-            var endTag = '</' + tagText + '>';
-            var start = textarea.prop('selectionStart');
-            var end = textarea.prop('selectionEnd');
-            var text = textarea.val();
-            var textBefore = text.substring(0, start);
-            var textInBetween;
-            var textAfter;
-            if (start === end) {
-                textAfter = text.substring(start, text.length);
-                textarea.val(textBefore + startTag + endTag + textAfter);
-            } else {
-                textInBetween = text.substring(start, end);
-                textAfter = text.substring(end, text.length);
-                textarea.val(textBefore + startTag + textInBetween + endTag + textAfter);
-            }
-            textarea.caret(startTag.length + text.length);
-        }
-    }
-
-    function keypadSlideLink(scope, elem, attrs) {
-        scope.$on('event:toggle', function () {
-            elem.animate({
-                width: 'toggle'
-            });
-        });
-        scope.$on('event:close', function () {
-            if (elem.css('display') == 'block') {
-                elem.animate({
-                    width: 'toggle'
-                });
-            }
-        });
-    }
-}
