--- app/modules/transcribe/overlay/keypad-slide.html
@@ -1,28 +1,0 @@
-<div class="keypad-slide" aria-label="Keypad">
-
-    <div class="heading">
-        <h3 class="title">Abbreviations</h3>
-    </div>
-
-    <div class="body">
-        <div class="buttons">
-            <div class="editing" role="group">
-                <div ng-repeat="abbreviation in abbreviations" data="abbreviation" ng-click="addTag(abbreviation.tag)" abbreviations-button></div>
-            </div>
-        </div>
-    </div>
-    <div class="heading">
-        <h3 class="title">TEI tags</h3>
-    </div>
-    <div class="body">
-
-        <div class="buttons">
-            <div class="editing" role="group">
-                <button class="button -tag" ng-bind="tag.name" ng-click="addTag(tag.tag)" ng-repeat="tag in tags"></button>
-            </div>
-
-        </div>
-
-    </div>
-
-</div>
