+++ app/modules/transcribe/overlay/overlay.html
@@ -3,6 +3,5 @@
     <div context-menu></div>
     <div transcribe-dialog></div>
     <div graphic-dialog></div>
-    <div keypad-slide></div>
     <div out-of-data></div>
 </div>
