+++ app/styles/components/graphic-dialog.styl
@@ -27,7 +27,6 @@
         line-height: 23px
 
     > .heading > .close
-        cursor: pointer
         position: relative
 
     > .heading > .close > .icon
