+++ app/styles/components/transcribe-dialog.styl
@@ -28,9 +28,7 @@
         line-height: 23px
 
     > .heading > .close
-        cursor: pointer
         position: relative
-        right: -500px
 
     > .heading > .close > .icon
         @extend .glyphicon
@@ -50,6 +48,7 @@
     > .body > .buttons
         display: flex
         justify-content: space-between
+        margin-bottom: 15px
 
     > .body > .buttons .button
         @extend .btn
